Manually export package.json to support introspective tools
Tools such as Components.js and react-native require introspection of
the package.json file. Since the new export syntax does not imply
package.json resolution yet (nodejs/node#33460), this has to be exported
explicitly. Ideally, this would be something that is done implicitly by
Node.js, but this is something that is still being discussed at
nodejs/node#33460. So until then, this has to be added explictly to
avoid breaking other tooling.

Related to uuidjs/uuid#444